# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GeekThrifts ecommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  isActive: boolean;
  productCount: number;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  isActive: boolean;
}

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  imageUrl?: string | null;
  categoryId: number;
  categoryName: string;
  sizes: string[];
  stock: number;
  isActive: boolean;
  isFeatured: boolean;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  description?: string | null;
  price: number;
  imageUrl?: string | null;
  categoryId: number;
  sizes: string[];
  stock: number;
  isActive: boolean;
  isFeatured: boolean;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  size: string;
  price: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  delivered: "delivered",
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  customerCity: string;
  notes?: string | null;
  status: OrderStatus;
  totalAmount: number;
  items: OrderItem[];
  createdAt: string;
}

export type CreateOrderBodyItemsItem = {
  productId: number;
  quantity: number;
  size: string;
};

export interface CreateOrderBody {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  customerCity: string;
  notes?: string | null;
  items: CreateOrderBodyItemsItem[];
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  delivered: "delivered",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  confirmedOrders: number;
  deliveredOrders: number;
  cancelledOrders: number;
  totalRevenue: number;
  totalProducts: number;
  totalCategories: number;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  delivered: "delivered",
} as const;