# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GeekThrifts ecommerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: categories
    description: Product categories
  - name: products
    description: Products
  - name: orders
    description: Orders
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, confirmed, cancelled, delivered]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by id (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        productCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, slug, isActive, productCount, createdAt]

    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [name, slug, isActive]

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
        sizes:
          type: array
          items:
            type: string
        stock:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          type: string
      required: [id, name, price, categoryId, categoryName, sizes, stock, isActive, isFeatured, createdAt]

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        categoryId:
          type: integer
        sizes:
          type: array
          items:
            type: string
        stock:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
      required: [name, price, categoryId, sizes, stock, isActive, isFeatured]

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        size:
          type: string
        price:
          type: number
      required: [productId, productName, quantity, size, price]

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        customerCity:
          type: string
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, cancelled, delivered]
        totalAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
      required: [id, customerName, customerPhone, customerAddress, customerCity, status, totalAmount, items, createdAt]

    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        customerCity:
          type: string
        notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              size:
                type: string
            required: [productId, quantity, size]
      required: [customerName, customerPhone, customerAddress, customerCity, items]

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, delivered]
      required: [status]

    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required: [success, token]

    AdminStats:
      type: object
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        confirmedOrders:
          type: integer
        deliveredOrders:
          type: integer
        cancelledOrders:
          type: integer
        totalRevenue:
          type: number
        totalProducts:
          type: integer
        totalCategories:
          type: integer
      required: [totalOrders, pendingOrders, confirmedOrders, deliveredOrders, cancelledOrders, totalRevenue, totalProducts, totalCategories]