# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GeekThrifts ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyItemsItem } from "./createOrderBodyItemsItem";

export interface CreateOrderBody {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  customerCity: string;
  notes?: string | null;
  items: CreateOrderBodyItemsItem[];
}