# lib/api-zod/src/generated/types/updateOrderStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GeekThrifts ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateOrderStatusBodyStatus } from "./updateOrderStatusBodyStatus";

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}